Clean-up of repo and near-final notebook
More clean-up and beautification of figures and saving them to a directory in the repo.

Removed some outdated analysis code, including some messing around in R, that we no longer need for the main analysis.

Changed the way data is loaded so as to avoid pulling the full dataset from the cloud for every re-run. Dataset now includes a CSV of just the columns we need from the raw scrape with a link to the full raw data.