Updated Analysis
This includes updates to the analysis following peer-review including calculating some more descriptive stats and adding some new Figures. New Figures are all saved in the appropriate directory.

Putting the entire analysis into a single Notebook was getting very difficult to work with so I've broken out each individual aspect of the analysis to its own notebook. Datasets that need to be shared between notebooks are noted and saved in the "Data" directory. All Notebooks should run to re-produce all numbers and figures as is. Certain Figures (2&4) were combined outside of the Notebook for the paper.

A .py file for the scraper for collecting data from the results section is also now included in the repo.